LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/container-units-animation-expected.txt


FAIL Animation using cqw unit assert_equals: expected "60px" but got "240px"
FAIL Animation using cqw unit responds to changing container size assert_equals: expected "60px" but got "240px"
FAIL Animation using cqh unit assert_equals: expected "60px" but got "180px"
FAIL Animation using cqh unit responds to changing container size assert_equals: expected "60px" but got "180px"
FAIL Animation using cqi unit assert_equals: expected "60px" but got "240px"
FAIL Animation using cqi unit responds to changing container size assert_equals: expected "60px" but got "240px"
FAIL Animation using cqb unit assert_equals: expected "60px" but got "180px"
FAIL Animation using cqb unit responds to changing container size assert_equals: expected "60px" but got "180px"
FAIL Animation using cqmin unit assert_equals: expected "60px" but got "180px"
FAIL Animation using cqmin unit responds to changing container size assert_equals: expected "60px" but got "180px"
FAIL Animation using cqmax unit assert_equals: expected "60px" but got "240px"
FAIL Animation using cqmax unit responds to changing container size assert_equals: expected "60px" but got "240px"
